requirements.txt
# Core dependencies
streamlit
python-dotenv
phidata
groq
# Search tools dependencies
duckduckgo-search
google-search-results
googlesearch-python
